Expose (private) API for creating QtRO types
The intent is to better support Python bindings.

This adds RepParser to the main source of RemoteObjects, where before
the parsing of .rep files into code had been independent of the
RemoteObjects source.  We need to know about the AST output from the
parser to generate the dynamic classes.

We add qremoteobjectrepparser_p.h (and a corresponding c++ file) with
the APIs to generate QMetaObjects dynamically and register the types
we can dynamically.

The structs used for dynamic enum/gadget/class types were moved to a
new header file so they could be shared with Node's dynamic parsing
as well as the newly added API.

[ChangeLog] Add external/private interface to support bindings
(specifically Python) to generate dynamic classes from .rep files.

Pick-to: 6.9
Change-Id: Id5bef60ae7c6e178ed9b2f8391224d05eda54662
Reviewed-by: Michael Brasser <michael.brasser@live.com>